requirements.txt
Django==4.2.3
python-dotenv==0.21.0
pre-commit==3.3.2
flake8==6.0.0
black==23.3.0
isort==5.12.0
psycopg2-binary==2.9.6
django-referrer-policy==1.0
django-csp==3.7
gunicorn==20.1.0
whitenoise==6.4.0
django-dsfr==0.14.1
factory-boy==3.3.0